youtube_watch_later/requirements.txt
# Core dependencies
Flask==2.3.3
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
oauth2client==4.1.3
google-auth==2.22.0

# HTTP and SSL
urllib3==2.0.4
requests==2.31.0
httplib2==0.22.0

# Date handling
python-dateutil==2.8.2
isodate==0.6.1

# Environment and config
python-dotenv==1.0.0

# Logging
colorlog==6.7.0

# Threading and scheduling
schedule==1.2.0

# Utils
tqdm==4.65.0
cachetools==5.3.1

# Development tools
black==23.7.0
pylint==2.17.5
pytest==7.4.0
pytest-cov==4.1.0

# Optional but recommended
cryptography==41.0.3
pyOpenSSL==23.2.0